# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: EduVibe Archive API
  version: 1.0.0
  description: Phase 1 public catalog reads and mock registration contract.
servers:
  - url: /api/v1
paths:
...
  /meta:
    get:
      operationId: getMeta
      summary: Read public catalog metadata
      description: >-
        Headerless metadata discovery is independent of authentication. If any
        connection header is supplied, all three must be canonical; otherwise 422.
      security: []
      parameters:
        - name: X-EduVibe-Flow-Id
          in: header
          required: false
          schema: { type: string }
        - name: X-EduVibe-Auth-Revision
          in: header
          required: false
          schema: { type: string }
        - name: X-EduVibe-Session-Generation
          in: header
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Public catalog metadata and feature availability.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meta"
        "400":
          $ref: "#/components/responses/ServiceError"
        "422":
          $ref: "#/components/responses/ServiceError"
        "503":
          $ref: "#/components/responses/ServiceError"
  /apps:
    get:
      operationId: listPublicApps
      summary: List public archive apps
      description: >-
        Returns publicly accessible apps ordered by created_at descending, then
        id descending. Subject facets describe the full accessible set and do
        not depend on search, grade, or page offset. Reject repeated and unknown
        query parameters.
      security: []
      parameters:
        - name: X-EduVibe-Flow-Id
          in: header
          required: false
          schema: { type: string }
        - name: X-EduVibe-Auth-Revision
          in: header
          required: false
          schema: { type: string }
        - name: X-EduVibe-Session-Generation
          in: header
          required: false
          schema: { type: string }
        - name: q
          in: query
          description: >-
            Trim, normalize to NFC, then apply Unicode default case folding.
            Count the post-folded value in Unicode code points; reject values
            longer than 100 without truncation. Search name, author nickname,
            and description as separate substring fields. Treat percent and
            underscore literally; do not search login IDs or contact fields.
          schema: { type: string }
        - name: subject
          in: query
          description: Empty omits the filter; otherwise use one catalog value.
          schema:
            anyOf:
              - type: string
                enum: [""]
              - $ref: "#/components/schemas/Subject"
        - name: grade
          in: query
          description: Empty omits the filter; otherwise use one catalog value.
          schema:
            anyOf:
              - type: string
                enum: [""]
              - $ref: "#/components/schemas/Grade"
        - name: limit
          in: query
          description: Page size; defaults to 24 and cannot exceed 100.
          schema: { type: integer, minimum: 1, maximum: 100, default: 24 }
        - name: offset
          in: query
          description: Zero-based offset. A valid offset past the result count returns an empty page.
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: A page of public app cards.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppPage"
        "400":
          $ref: "#/components/responses/ServiceError"
        "409":
          $ref: "#/components/responses/ServiceError"
        "422":
          $ref: "#/components/responses/ServiceError"
        "503":
          $ref: "#/components/responses/ServiceError"
    post:
      operationId: createApp
      summary: Create an archive app with an issued operation key
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/AuthFlowId"
        - $ref: "#/components/parameters/AuthRevision"
        - $ref: "#/components/parameters/SessionGeneration"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AppInput" }
      responses:
        "201":
          description: The persisted app and its current owner.
          headers:
            Cache-Control: { schema: { type: string, const: no-store } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppDetailResponse" }
        "400": { $ref: "#/components/responses/ServiceError" }
        "401": { $ref: "#/components/responses/ServiceError" }
        "403": { $ref: "#/components/responses/ServiceError" }
        "404": { $ref: "#/components/responses/ServiceError" }
        "409": { $ref: "#/components/responses/ServiceError" }
        "410": { $ref: "#/components/responses/ServiceError" }
        "413": { $ref: "#/components/responses/ServiceError" }
        "422": { $ref: "#/components/responses/ServiceError" }
        "429": { $ref: "#/components/responses/ServiceError" }
        "503": { $ref: "#/components/responses/ServiceError" }
  /apps/{id}:
    get:
      operationId: getApp
      summary: Read an archive app available to the current member
      security:
        - {}
        - SessionCookie: []
      parameters:
        - name: X-EduVibe-Flow-Id
          in: header
          required: false
          schema: { type: string }
        - name: X-EduVibe-Auth-Revision
          in: header
          required: false
          schema: { type: string }
        - name: X-EduVibe-Session-Generation
          in: header
          required: false
          schema: { type: string }
        - name: id
          in: path
          required: true
          description: >-
            Canonical 8-4-4-4-12 hexadecimal UUID with version 1-8 and variant
            8/9/a/b. Hex is case-insensitive and normalized to lowercase.
            Other formats return 404 NOT_FOUND; query errors take precedence.
          schema:
            type: string
            format: uuid
            pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
            minLength: 36
            maxLength: 36
      responses:
        "200":
          description: >-
            Public detail, or private detail authorized for the current approved
            full owner or full administrator. Headerless requests ignore cookies.
            Supplied context is all-or-none and canonical; stale or pending context
            is rejected without replay. Missing and unauthorized detail share 404.
            Member-bound responses include verified context and are never cached.
          headers:
            Cache-Control:
              description: Present on member-bound responses only.
              schema: { type: string, const: "private, no-store" }
            X-EduVibe-Flow-Id:
              description: Verified flow ID, present on member-bound responses only.
              schema: { type: string, format: uuid }
            X-EduVibe-Auth-Revision:
              description: Verified revision, present on member-bound responses only.
              schema: { type: string }
            X-EduVibe-Session-Generation:
              description: Verified generation, present on member-bound responses only.
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppDetailResponse"
        "400":
          $ref: "#/components/responses/ServiceError"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/ServiceError"
        "403":
          $ref: "#/components/responses/ServiceError"
        "409":
          $ref: "#/components/responses/ServiceError"
        "422":
          $ref: "#/components/responses/ServiceError"
        "503":
          $ref: "#/components/responses/ServiceError"
    patch:
      operationId: updateApp
      summary: Update an owned archive app at its expected version
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/AuthFlowId"
        - $ref: "#/components/parameters/AuthRevision"
        - $ref: "#/components/parameters/SessionGeneration"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateAppInput" }
      responses:
        "200":
          description: The persisted app at its new version.
          headers:
            Cache-Control: { schema: { type: string, const: no-store } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppDetailResponse" }
        "400": { $ref: "#/components/responses/ServiceError" }
        "401": { $ref: "#/components/responses/ServiceError" }
        "403": { $ref: "#/components/responses/ServiceError" }
        "404": { $ref: "#/components/responses/ServiceError" }
        "409": { $ref: "#/components/responses/ServiceError" }
        "410": { $ref: "#/components/responses/ServiceError" }
        "413": { $ref: "#/components/responses/ServiceError" }
        "422": { $ref: "#/components/responses/ServiceError" }
        "429": { $ref: "#/components/responses/ServiceError" }
        "503": { $ref: "#/components/responses/ServiceError" }
    delete:
      operationId: deleteApp
      summary: Delete an owned archive app at its expected version
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/AuthFlowId"
        - $ref: "#/components/parameters/AuthRevision"
        - $ref: "#/components/parameters/SessionGeneration"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeleteAppInput" }
      responses:
        "204":
          description: The app and its related health result and jobs were deleted.
        "400": { $ref: "#/components/responses/ServiceError" }
        "401": { $ref: "#/components/responses/ServiceError" }
        "403": { $ref: "#/components/responses/ServiceError" }
        "404": { $ref: "#/components/responses/ServiceError" }
        "409": { $ref: "#/components/responses/ServiceError" }
        "410": { $ref: "#/components/responses/ServiceError" }
        "413": { $ref: "#/components/responses/ServiceError" }
        "422": { $ref: "#/components/responses/ServiceError" }
        "429": { $ref: "#/components/responses/ServiceError" }
        "503":
          $ref: "#/components/responses/ServiceError"
          description: >-
            Includes DELETION_CONFIRMATION_PENDING when the app deletion is
            committed but its related deletion record is not yet confirmed.
            Check the existing operation key; do not submit a new key.
...
  /admin/apps:
    get:
      operationId: listAdminApps
      summary: List all public and private apps for the administrator monitor
      description: >-
        Returns only the app name, URL, visibility, theme, versions, creation
        time, author nickname, and connection result summary. Prompts,
        descriptions, and member contact fields are never included. Items sort
        by creation time descending and app ID descending.
      security:
        - SessionCookie: []
      parameters:
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 100, default: 24 },
          }
        - {
            name: offset,
            in: query,
            schema: { type: integer, minimum: 0, default: 0 },
          }
        - $ref: "#/components/parameters/AuthFlowId"
        - $ref: "#/components/parameters/AuthRevision"
        - $ref: "#/components/parameters/SessionGeneration"
      responses:
        "200":
          description: A page of all authorized apps with only monitor fields.
          headers:
            Cache-Control: { schema: { type: string, const: no-store } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminAppPage" }
        "401": { $ref: "#/components/responses/ServiceError" }
        "403": { $ref: "#/components/responses/ServiceError" }
        "409": { $ref: "#/components/responses/ServiceError" }
        "503": { $ref: "#/components/responses/ServiceError" }
...
  /write-operations:
    post:
      operationId: createWriteOperation
      summary: Issue a write operation key without changing app or account data
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/AuthFlowId"
        - $ref: "#/components/parameters/AuthRevision"
        - $ref: "#/components/parameters/SessionGeneration"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateWriteOperation" }
      responses:
        "201":
          description: An unresolved key result with no echoed business input.
          headers:
            Cache-Control: { schema: { type: string, const: "private, no-store" } }
            X-EduVibe-Flow-Id: { schema: { type: string, format: uuid } }
            X-EduVibe-Auth-Revision: { schema: { type: string, pattern: "^(0|[1-9][0-9]*)$" } }
            X-EduVibe-Session-Generation: { schema: { type: string, pattern: "^(0|[1-9][0-9]*)$" } }
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ApprovalOperation"
                  - $ref: "#/components/schemas/PasswordResetOperation"
                  - $ref: "#/components/schemas/UserDeleteOperation"
                  - $ref: "#/components/schemas/AppWriteOperation"
        "400": { $ref: "#/components/responses/ServiceError" }
        "401": { $ref: "#/components/responses/ServiceError" }
        "403": { $ref: "#/components/responses/ServiceError" }
        "404": { $ref: "#/components/responses/ServiceError" }
        "409": { $ref: "#/components/responses/ServiceError" }
        "413": { $ref: "#/components/responses/ServiceError" }
        "422": { $ref: "#/components/responses/ServiceError" }
        "503": { $ref: "#/components/responses/ServiceError" }
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
...
    Origin:
      name: Origin
      in: header
      required: true
      schema: { type: string, format: uri }
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 1 }
    AuthFlowId:
      name: X-EduVibe-Flow-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
    AuthRevision:
      name: X-EduVibe-Auth-Revision
      in: header
      required: true
      schema: { type: string, pattern: "^(0|[1-9][0-9]*)$" }
    SessionGeneration:
      name: X-EduVibe-Session-Generation
      in: header
      required: true
      schema: { type: string, pattern: "^(0|[1-9][0-9]*)$" }
...
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: __Host-eduvibe_session_<flow_id>_<issued_seq>
      description: >-
        Production cookies use a fresh name per flow and issued sequence.
        Explicit HTTP localhost/loopback development names are
        eduvibe_session_dev_<flow_id>_<issued_seq>.
    RecoveryCookie:
      type: apiKey
      in: cookie
      name: __Host-eduvibe_recovery_<flow_id>_<issued_seq>
      description: >-
        The name is generated per flow. The placeholders stand for the request's
        X-EduVibe-Flow-Id and the server-issued recovery sequence.
  responses:
    NotFound:
      description: The app does not exist or is not accessible to the current member.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ServiceError:
      description: The request failed or the response could not be accepted.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    DeleteAppInput:
      type: object
      additionalProperties: false
      required: [expected_version]
      properties:
        expected_version:
          type: integer
          minimum: 1
          maximum: 9007199254740991
...
    UpdateAppInput:
      type: object
      additionalProperties: false
      required: [expected_version]
      minProperties: 2
      properties:
        expected_version:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        name: { type: string, minLength: 1, maxLength: 100 }
        url: { type: string, minLength: 1, maxLength: 2048, format: uri }
        prompt: { type: string, minLength: 1, maxLength: 30000 }
        description: { type: string, minLength: 1, maxLength: 20000 }
        subject: { $ref: "#/components/schemas/Subject" }
        grades:
          type: array
          minItems: 1
          uniqueItems: true
          items: { $ref: "#/components/schemas/Grade" }
        is_public: { type: boolean }
        theme_id: { type: string }
        stack_db: { type: [string, "null"], maxLength: 200 }
        stack_backend: { type: [string, "null"], maxLength: 200 }
        stack_frontend: { type: [string, "null"], maxLength: 200 }
        stack_hosting: { type: [string, "null"], maxLength: 200 }
...
    Subject:
      type: string
      enum: [수학, 과학, 영어, 역사, 국어, 사회, 정보, 기타]
    Grade:
      type: string
      enum: [초1, 초2, 초3, 초4, 초5, 초6, 중1, 중2, 중3, 고1, 고2, 고3]
...
    Meta:
      type: object
      additionalProperties: false
      required:
        [
          subjects,
          grades,
          themes,
          server_time,
          capabilities,
          support,
          initial_pending_days,
        ]
      properties:
        subjects:
          { type: array, items: { $ref: "#/components/schemas/Subject" } }
        grades: { type: array, items: { $ref: "#/components/schemas/Grade" } }
        themes:
          type: array
          items: { $ref: "#/components/schemas/Theme" }
        server_time: { type: string, format: date-time }
        capabilities: { $ref: "#/components/schemas/Capabilities" }
        support: { $ref: "#/components/schemas/Support" }
        initial_pending_days: { type: integer, minimum: 1 }
...
    Support:
      type: object
      additionalProperties: false
      required: [email, service_url, announcement_url]
      properties:
        email: { type: [string, "null"] }
        service_url: { type: [string, "null"], format: uri }
        announcement_url: { type: [string, "null"], format: uri }
...
    AppPage:
      type: object
      additionalProperties: false
      required: [items, pagination, server_time, facets]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/AppCard" } }
        pagination: { $ref: "#/components/schemas/Pagination" }
        server_time: { type: string, format: date-time }
        facets:
          type: object
          additionalProperties: false
          required: [subjects_in_use]
          properties:
            subjects_in_use:
              { type: array, items: { $ref: "#/components/schemas/Subject" } }
...
    AppDetailResponse:
      type: object
      additionalProperties: false
      required: [item, server_time]
      properties:
        item: { $ref: "#/components/schemas/AppDetail" }
        server_time: { type: string, format: date-time }
    AppInput:
      type: object
      additionalProperties: false
      required:
        [
          name,
          url,
          prompt,
          description,
          subject,
          grades,
          is_public,
          theme_id,
          stack_db,
          stack_backend,
          stack_frontend,
          stack_hosting,
        ]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        url: { type: string, minLength: 1, maxLength: 2048, format: uri }
        prompt: { type: string, minLength: 1, maxLength: 30000 }
        description: { type: string, minLength: 1, maxLength: 20000 }
        subject: { $ref: "#/components/schemas/Subject" }
        grades:
          type: array
          minItems: 1
          uniqueItems: true
          items: { $ref: "#/components/schemas/Grade" }
        is_public: { type: boolean }
        theme_id: { type: string }
        stack_db: { type: [string, "null"], maxLength: 200 }
        stack_backend: { type: [string, "null"], maxLength: 200 }
        stack_frontend: { type: [string, "null"], maxLength: 200 }
        stack_hosting: { type: [string, "null"], maxLength: 200 }